requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
numpy>=1.24.0
Pillow>=10.0.0
gspread>=5.12.0
google-auth>=2.23.0